Added one more check for admin user (user name needs to be 'admin') when deleting datasets.

+++ src/services/gov/epa/emissions/framework/services/data/DataServiceImpl.java
@@ -328,7 +328,7 @@ public class DataServiceImpl implements DataService {
         DbServer dbServer = DbServerFactory.get().getDbServer();
 
         try {
-            if (user.isAdmin())
+            if (user.getUsername().equals("admin") && user.isAdmin())
                 dao.removeEmptyDatasets(user, dbServer, session);
             
             List<EmfDataset> list = dao.deletedDatasets(user, session);
